+++ MillimanAccessPortal/MapTests/AuthorizedContentControllerTests.cs
@@ -43,7 +43,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -85,7 +86,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -132,7 +134,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -177,7 +180,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -187,8 +191,8 @@ namespace MapTests
 
             // Add a file to the root content item and a content url to the selection group
             string FileName = "CCR_0273ZDM_New_Reduction_Script.qvw";
-            string TestFileSourcePath = Path.Combine(@"\\indy-syn01\prm_test\Sample Data", FileName);
-            string TestFileTargetPath = Path.Combine(@"\\indy-syn01\prm_test\ContentRoot", TestUtil.MakeTestGuid(1).ToString(), FileName);
+            string TestFileSourcePath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\Sample Data", FileName);
+            string TestFileTargetPath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\ContentRoot", TestUtil.MakeTestGuid(1).ToString(), FileName);
             File.Copy(TestFileSourcePath, TestFileTargetPath, true);
             SelectionGroup ThisGroup = TestResources.DbContextObject.SelectionGroup.Single(sg => sg.Id == TestUtil.MakeTestGuid(1));
             RootContentItem ThisItem = TestResources.DbContextObject.RootContentItem.FirstOrDefault(rci => rci.Id == TestUtil.MakeTestGuid(1));
@@ -240,7 +244,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -250,8 +255,8 @@ namespace MapTests
 
             // Add a file to the root content item and a content url to the selection group
             string FileName = "CCR_0273ZDM_New_Reduction_Script.qvw";
-            string TestFileSourcePath = Path.Combine(@"\\indy-syn01\prm_test\Sample Data", FileName);
-            string TestFileTargetPath = Path.Combine(@"\\indy-syn01\prm_test\ContentRoot", TestUtil.MakeTestGuid(1).ToString(), FileName);
+            string TestFileSourcePath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\Sample Data", FileName);
+            string TestFileTargetPath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\ContentRoot", TestUtil.MakeTestGuid(1).ToString(), FileName);
             File.Copy(TestFileSourcePath, TestFileTargetPath, true);
             SelectionGroup ThisGroup = TestResources.DbContextObject.SelectionGroup.Single(sg => sg.Id == TestUtil.MakeTestGuid(1));
             RootContentItem ThisItem = TestResources.DbContextObject.RootContentItem.FirstOrDefault(rci => rci.Id == TestUtil.MakeTestGuid(1));
@@ -297,7 +302,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -339,7 +345,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -384,7 +391,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -392,8 +400,8 @@ namespace MapTests
             // Following throws if dependency failed to create or specified user is not in the data. Use try/catch to prevent failure for this cause
             sut.ControllerContext = TestInitialization.GenerateControllerContext(userName: TestResources.DbContextObject.ApplicationUser.Where(u => u.UserName == "test1").First().UserName);
 
-            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01\prm_test\Sample Data", "IHopeSo.pdf");
-            string UserGuideTestPath = Path.Combine(@"\\indy-syn01\prm_test\ContentRoot", purpose + ".pdf");
+            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\Sample Data", "IHopeSo.pdf");
+            string UserGuideTestPath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\ContentRoot", purpose + ".pdf");
             File.Copy(UserGuideSourcePath, UserGuideTestPath, true);
             RootContentItem ThisItem = TestResources.DbContextObject.RootContentItem.Single(rci => rci.Id == TestUtil.MakeTestGuid(1));
             ThisItem.ContentFilesList = new List<MapDbContextLib.Models.ContentRelatedFile>
@@ -446,7 +454,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -454,8 +463,8 @@ namespace MapTests
             // Following throws if dependency failed to create or specified user is not in the data. Use try/catch to prevent failure for this cause
             sut.ControllerContext = TestInitialization.GenerateControllerContext(userName: TestResources.DbContextObject.ApplicationUser.Where(u => u.UserName == "test1").First().UserName);
             string purpose = "UserGuide";
-            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01\prm_test\Sample Data", "IHopeSo.pdf");
-            string UserGuideTestPath = Path.Combine(@"\\indy-syn01\prm_test\ContentRoot", purpose + ".pdf");
+            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\Sample Data", "IHopeSo.pdf");
+            string UserGuideTestPath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\ContentRoot", purpose + ".pdf");
             File.Copy(UserGuideSourcePath, UserGuideTestPath, true);
             RootContentItem ThisItem = TestResources.DbContextObject.RootContentItem.Single(rci => rci.Id == TestUtil.MakeTestGuid(1));
             ThisItem.ContentFilesList = new List<MapDbContextLib.Models.ContentRelatedFile>
@@ -506,7 +515,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -515,8 +525,8 @@ namespace MapTests
             sut.ControllerContext = TestInitialization.GenerateControllerContext(userName: TestResources.DbContextObject.ApplicationUser.Where(u => u.UserName == "test1").First().UserName);
 
             string purpose = "UserGuide";
-            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01\prm_test\Sample Data", "IHopeSo.pdf");
-            string UserGuideTestPath = Path.Combine(@"\\indy-syn01\prm_test\ContentRoot", purpose + ".pdf");
+            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\Sample Data", "IHopeSo.pdf");
+            string UserGuideTestPath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\ContentRoot", purpose + ".pdf");
             File.Copy(UserGuideSourcePath, UserGuideTestPath, true);
             RootContentItem ThisItem = TestResources.DbContextObject.RootContentItem.Single(rci => rci.Id == TestUtil.MakeTestGuid(1));
             ThisItem.ContentFilesList = new List<MapDbContextLib.Models.ContentRelatedFile>
@@ -566,7 +576,8 @@ namespace MapTests
                 TestResources.QvConfig,
                 TestResources.QueriesObj,
                 TestResources.UserManagerObject,
-                TestResources.ConfigurationObject);
+                TestResources.ConfigurationObject,
+                TestResources.PowerBiConfig);
 
             // For illustration only, the same result comes from either of the following techniques:
             // This one should never throw even if the user name is not in the context data
@@ -575,8 +586,8 @@ namespace MapTests
             sut.ControllerContext = TestInitialization.GenerateControllerContext(userName: TestResources.DbContextObject.ApplicationUser.Where(u => u.UserName == "test1").First().UserName);
 
             string purpose = "UserGuide";
-            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01\prm_test\Sample Data", "IHopeSo.pdf");
-            string UserGuideTestPath = Path.Combine(@"\\indy-syn01\prm_test\ContentRoot", purpose + ".pdf");
+            string UserGuideSourcePath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\Sample Data", "IHopeSo.pdf");
+            string UserGuideTestPath = Path.Combine(@"\\indy-syn01.milliman.com\prm_test\ContentRoot", purpose + ".pdf");
             File.Copy(UserGuideSourcePath, UserGuideTestPath, true);
             RootContentItem ThisItem = TestResources.DbContextObject.RootContentItem.Single(rci => rci.Id == TestUtil.MakeTestGuid(1));
             ThisItem.ContentFilesList = new List<MapDbContextLib.Models.ContentRelatedFile>