+++ MillimanAccessPortal/MillimanAccessPortal/Services/QueuedGoLiveTaskHostedService.cs
@@ -18,6 +18,7 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using MillimanAccessPortal.Models.ContentPublishing;
 using MillimanAccessPortal.Services;
+using PowerBiLib;
 using QlikviewLib;
 using Serilog;
 using System;
@@ -44,39 +45,49 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
     {
         while (!cancellationToken.IsCancellationRequested)
         {
-            // Retrieve the relevant data to finalize the goLive
-            var goLiveViewModel = await TaskQueue.DequeueAsync(cancellationToken);
-
-            using (var scope = Services.CreateScope())
+            try
             {
-                var auditLogger = scope.ServiceProvider.GetRequiredService<IAuditLogger>();
-                var configuration = scope.ServiceProvider.GetRequiredService<IConfiguration>();
-                var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-                // Other content types may require their own services
-                var qlikviewConfig = scope.ServiceProvider.GetRequiredService<IOptions<QlikviewConfig>>().Value;
+                // Retrieve the relevant data to finalize the goLive
+                GoLiveViewModel goLiveViewModel = await TaskQueue.DequeueAsync(cancellationToken);
 
-                try
+                using (var scope = Services.CreateScope())
                 {
-                    await ProcessGoLive(
-                        goLiveViewModel,
-                        dbContext,
-                        auditLogger,
-                        configuration,
-                        qlikviewConfig);
+                    var auditLogger = scope.ServiceProvider.GetRequiredService<IAuditLogger>();
+                    var configuration = scope.ServiceProvider.GetRequiredService<IConfiguration>();
+                    var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+                    ContentTypeEnum contentType = dbContext.RootContentItem
+                                                           .Where(c => c.Id == goLiveViewModel.RootContentItemId)
+                                                           .Select(c => c.ContentType.TypeEnum)
+                                                           .Single();
+
+                    object typeSpecificConfig = default;
+                    switch (contentType)
+                    {
+                        case ContentTypeEnum.Qlikview:
+                            typeSpecificConfig = scope.ServiceProvider.GetRequiredService<IOptions<QlikviewConfig>>().Value;
+                            break;
+                        case ContentTypeEnum.PowerBi:
+                            typeSpecificConfig = scope.ServiceProvider.GetRequiredService<IOptions<PowerBiConfig>>().Value;
+                            break;
+                    }
+
+                    await ProcessGoLive(goLiveViewModel, dbContext, auditLogger, configuration, typeSpecificConfig);
                 }
-                catch (Exception e)
-                {
-                    Log.Error(e, "In QueueGoLiveTaskHostedService.ExecuteAsync");
-                }
+            }
+            catch (Exception e)
+            {
+                Log.Error(e, "In QueueGoLiveTaskHostedService.ExecuteAsync");
             }
         }
     }
 
     private async Task ProcessGoLive(
-        GoLiveViewModel goLiveViewModel, ApplicationDbContext dbContext, IAuditLogger auditLogger,
-        IConfiguration configuration, QlikviewConfig qlikviewConfig)
+        GoLiveViewModel goLiveViewModel, 
+        ApplicationDbContext dbContext, 
+        IAuditLogger auditLogger,
+        IConfiguration configuration, 
+        object typeSpecificConfig)
     {
-        #region Checksum verification
         var publicationRequest = goLiveViewModel == null
             ? null
             : dbContext.ContentPublicationRequest
@@ -87,6 +98,7 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 .Where(r => r.RootContentItemId == goLiveViewModel.RootContentItemId)
                 .SingleOrDefault(r => r.RequestStatus == PublicationStatus.Confirming);
 
+        #region Validation
         if (publicationRequest?.RootContentItem == null || publicationRequest?.ApplicationUser == null)
         {
             Log.Error(
@@ -96,6 +108,7 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 $"or related content item {publicationRequest?.RootContentItemId} not found");
             return;
         }
+        #endregion
 
         var LiveHierarchy = new ContentReductionHierarchy<ReductionFieldValue>
         {
@@ -136,13 +149,9 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 catch (InvalidOperationException)
                 {
                     Log.Error($"" +
-                        "In QueueGoLiveTaskHostedService.ExecuteAsync action: " +
-                        "expected one reduction task for each non-master selection group, " +
-                        $"failed for selection group {relatedSelectionGroup.Id}, " +
-                        "aborting");
-                    await FailGoLive(dbContext, publicationRequest,
-                        $"Expected 1 reduction task related to SelectionGroup {relatedSelectionGroup.Id}, " +
-                        "cannot complete this go-live request.");
+                        "In QueueGoLiveTaskHostedService.ExecuteAsync action: expected `Single` reduction task for each non-master selection group, " +
+                        $"failed for selection group {relatedSelectionGroup.Id}, aborting");
+                    await FailGoLiveAsync(dbContext, publicationRequest, $"Expected 1 reduction task related to SelectionGroup {relatedSelectionGroup.Id}, cannot complete this go-live request.");
                     return;
                 }
 
@@ -154,14 +163,12 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                     if (currentChecksum != ThisTask.ReducedContentChecksum.ToLower())
                     {
                         Log.Error($"In QueueGoLiveTaskHostedService.ExecuteAsync action: " +
-                            "for selection group {relatedSelectionGroup.Id}, " +
-                            "reduced content file {ThisTask.ResultFilePath} failed checksum validation, " +
-                            "aborting");
+                            $"for selection group {relatedSelectionGroup.Id}, " +
+                            $"reduced content file {ThisTask.ResultFilePath} failed checksum validation, aborting");
                         auditLogger.Log(AuditEventType.GoLiveValidationFailed.ToEvent(
                             publicationRequest.RootContentItem, publicationRequest));
-                        await FailGoLive(dbContext, publicationRequest,
-                            $"Reduced content file failed integrity check, " +
-                            "cannot complete the go-live request.");
+                        await FailGoLiveAsync(dbContext, publicationRequest,
+                            $"Reduced content file failed integrity check, cannot complete the go-live request.");
                         return;
                     }
                 }
@@ -183,11 +190,10 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                     "aborting");
                 auditLogger.Log(AuditEventType.GoLiveValidationFailed.ToEvent(
                     publicationRequest.RootContentItem, publicationRequest));
-                await FailGoLive(dbContext, publicationRequest, "File integrity validation failed");
+                await FailGoLiveAsync(dbContext, publicationRequest, "File integrity validation failed");
                 return;
             }
         }
-        #endregion
 
         #region Go live
         List<Action> successActionList = new List<Action>();
@@ -205,7 +211,6 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 foreach (ContentPublicationRequest PreviousLiveRequest in previousLiveRequests)
                 {
                     PreviousLiveRequest.RequestStatus = PublicationStatus.Replaced;
-                    dbContext.ContentPublicationRequest.Update(PreviousLiveRequest);
                 }
                 publicationRequest.RequestStatus = PublicationStatus.Confirmed;
 
@@ -218,16 +223,14 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                     foreach (ContentReductionTask PreviousLiveTask in previousLiveTasks)
                     {
                         PreviousLiveTask.ReductionStatus = ReductionStatusEnum.Replaced;
-                        dbContext.ContentReductionTask.Update(PreviousLiveTask);
                     }
                     relatedReductionTasks.ForEach(t => t.ReductionStatus = ReductionStatusEnum.Live);
                 }
 
                 //1.3  HierarchyFieldValue due to hierarchy changes
-                //1.3.1  If this is first publication for this root content item, add the fields to db
-                //       and to LiveHierarchy to help identify all values as new
+                //1.3.1  If this is first publication for this RootContentItem add the fields to db and to LiveHierarchy (not values to LiveHierarchy)
                 if (LiveHierarchy.Fields.Count == 0)
-                {  // This must be first time publication, need to insert the fields.  Values are handled below
+                {  // This must be first time publication, need to insert the fields.  Field values are handled later
                     NewHierarchy.Fields.ForEach(f =>
                     {
                         HierarchyField NewField = new HierarchyField
@@ -239,7 +242,6 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                             StructureType = f.StructureType,
                         };
                         dbContext.HierarchyField.Add(NewField);
-                        dbContext.SaveChanges();
 
                         LiveHierarchy.Fields.Add(new ReductionField<ReductionFieldValue>
                         {
@@ -252,6 +254,8 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                         });
                     });
                 }
+                dbContext.SaveChanges();
+
                 //1.3.2  Add/Remove field values based on value list differences between new/old
                 foreach (var NewHierarchyField in NewHierarchy.Fields)
                 {
@@ -301,54 +305,84 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 List<ContentRelatedFile> UpdatedContentFilesList = publicationRequest.RootContentItem.ContentFilesList;
                 foreach (ContentRelatedFile Crf in publicationRequest.LiveReadyFilesObj)
                 {
-                    // This assignment defines the live file name
                     string TargetFileName = ContentTypeSpecificApiBase.GenerateContentFileName(
                         Crf.FilePurpose, Path.GetExtension(Crf.FullPath), goLiveViewModel.RootContentItemId);
-                    string TargetFilePath = Path.Combine(configuration.GetSection("Storage")["ContentItemRootPath"],
-                                                         goLiveViewModel.RootContentItemId.ToString(),
-                                                         TargetFileName);
 
-                    // Move any existing live file of this name to backed up name
-                    if (File.Exists(TargetFilePath))
+                    // special treatment for powerbi content file (no live content file persists in MAP)
+                    if (Crf.FilePurpose.Equals("mastercontent", StringComparison.OrdinalIgnoreCase) && 
+                        publicationRequest.RootContentItem.ContentType.TypeEnum == ContentTypeEnum.PowerBi)
                     {
-                        string BackupFilePath = TargetFilePath + ".bak";
-                        if (File.Exists(BackupFilePath))
-                        {
-                            File.Delete(BackupFilePath);
-                        }
-                        File.Move(TargetFilePath, BackupFilePath);
+                        PowerBiContentItemProperties typeSpecificProperties = publicationRequest.RootContentItem.TypeSpecificDetailObject as PowerBiContentItemProperties;
 
-                        successActionList.Add(new Action(() => {
-                            File.Delete(BackupFilePath);
-                        }));
-                        failureRecoveryActionList.Add(new Action(() => {
-                            if (File.Exists(TargetFilePath))
-                            {
-                                File.Delete(TargetFilePath);
-                            }
-                            File.Move(BackupFilePath, TargetFilePath);
-                        }));
+                        failureRecoveryActionList.Add(() => {
+                            publicationRequest.RootContentItem.TypeSpecificDetailObject = typeSpecificProperties;
+                            dbContext.SaveChanges();
+                        });
+                        string reportIdToBeDeleted = typeSpecificProperties.LiveReportId;
+                        successActionList.Add(async () => {
+                            PowerBiLibApi powerBiApi = await new PowerBiLibApi(typeSpecificConfig as PowerBiConfig).InitializeAsync();
+                            bool deleteSucceeded = await powerBiApi.DeleteReportAsync(reportIdToBeDeleted);
+                        });
+
+                        typeSpecificProperties.LiveEmbedUrl = typeSpecificProperties.PreviewEmbedUrl;
+                        typeSpecificProperties.LiveReportId = typeSpecificProperties.PreviewReportId;
+                        typeSpecificProperties.LiveWorkspaceId = typeSpecificProperties.PreviewWorkspaceId;
+                        typeSpecificProperties.PreviewEmbedUrl = null;
+                        typeSpecificProperties.PreviewReportId = null; ;
+                        typeSpecificProperties.PreviewWorkspaceId = null;
+
+                        publicationRequest.RootContentItem.TypeSpecificDetailObject = typeSpecificProperties;
+
                     }
-
-                    // Can move since files are on the same volume
-                    File.Move(Crf.FullPath, TargetFilePath);
-
-                    failureRecoveryActionList.Insert(0, new Action(() => {  // This one must run before the one in the if block above
-                        if (File.Exists(Crf.FullPath))
-                        {
-                            File.Delete(Crf.FullPath);
-                        }
-                        File.Move(TargetFilePath, Crf.FullPath);
-                    }));
-
-                    UpdatedContentFilesList.RemoveAll(f => f.FilePurpose.ToLower() == Crf.FilePurpose.ToLower());
-                    UpdatedContentFilesList.Add(new ContentRelatedFile
+                    else
                     {
-                        FilePurpose = Crf.FilePurpose,
-                        FullPath = TargetFilePath,
-                        Checksum = Crf.Checksum,
-                        FileOriginalName = Crf.FileOriginalName,
-                    });
+                        // This assignment defines the live file name
+                        string TargetFilePath = Path.Combine(configuration.GetSection("Storage")["ContentItemRootPath"],
+                                                             goLiveViewModel.RootContentItemId.ToString(),
+                                                             TargetFileName);
+
+                        // Move any existing live file of this name to backed up name
+                        if (File.Exists(TargetFilePath))
+                        {
+                            string BackupFilePath = TargetFilePath + ".bak";
+                            if (File.Exists(BackupFilePath))
+                            {
+                                File.Delete(BackupFilePath);
+                            }
+                            File.Move(TargetFilePath, BackupFilePath);
+
+                            successActionList.Add(new Action(() => {
+                                File.Delete(BackupFilePath);
+                            }));
+                            failureRecoveryActionList.Add(new Action(() => {
+                                if (File.Exists(TargetFilePath))
+                                {
+                                    File.Delete(TargetFilePath);
+                                }
+                                File.Move(BackupFilePath, TargetFilePath);
+                            }));
+                        }
+
+                        // Can move since files are on the same volume
+                        File.Move(Crf.FullPath, TargetFilePath);
+
+                        failureRecoveryActionList.Insert(0, new Action(() => {  // This one must run before the one in the if block above
+                            if (File.Exists(Crf.FullPath))
+                            {
+                                File.Delete(Crf.FullPath);
+                            }
+                            File.Move(TargetFilePath, Crf.FullPath);
+                        }));
+
+                        UpdatedContentFilesList.RemoveAll(f => f.FilePurpose.ToLower() == Crf.FilePurpose.ToLower());
+                        UpdatedContentFilesList.Add(new ContentRelatedFile
+                        {
+                            FilePurpose = Crf.FilePurpose,
+                            FullPath = TargetFilePath,
+                            Checksum = Crf.Checksum,
+                            FileOriginalName = Crf.FileOriginalName,
+                        });
+                    }
 
                     // Set content URL in each master SelectionGroup
                     if (Crf.FilePurpose.ToLower() == "mastercontent")
@@ -443,11 +477,10 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
                 switch (publicationRequest.RootContentItem.ContentType.TypeEnum)
                 {
                     case ContentTypeEnum.Qlikview:
-                        await new QlikviewLibApi()
-                            .AuthorizeUserDocumentsInFolder(
-                                goLiveViewModel.RootContentItemId.ToString(), qlikviewConfig);
+                        await new QlikviewLibApi(typeSpecificConfig as QlikviewConfig).AuthorizeUserDocumentsInFolderAsync(goLiveViewModel.RootContentItemId.ToString());
                         break;
 
+                    case ContentTypeEnum.PowerBi:
                     case ContentTypeEnum.Html:
                     case ContentTypeEnum.Pdf:
                     case ContentTypeEnum.FileDownload:
@@ -518,7 +551,7 @@ public class QueuedGoLiveTaskHostedService : BackgroundService
         #endregion
     }
 
-    private async Task FailGoLive(
+    private async Task FailGoLiveAsync(
         ApplicationDbContext dbContext, ContentPublicationRequest publicationRequest, string reason)
     {
         publicationRequest.RequestStatus = PublicationStatus.Error;